[Opt] Remove unnecessary PROJECT nodes from plan
On top of each subquery or view, there is a PROJECT node. It's
unnecessary and will interfere various optimizer rules, such as join
order and aggregate push-down/pull-up. This commit removes most of that
kind of PROJECT node, but some are retained to avoid intricate issues.
