+++ pkg/sql/plan/opt_misc.go
@@ -1,0 +1,477 @@
+// Copyright 2023 Matrix Origin
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package plan
+
+import "github.com/matrixorigin/matrixone/pkg/pb/plan"
+
+// removeSimpleProjections On top of each subquery or view it has a PROJECT node, which interrupts optimizer rules such as join order.
+func (builder *QueryBuilder) removeSimpleProjections(nodeID int32, parentType plan.Node_NodeType) (int32, map[[2]int32]*plan.Expr) {
+	projMap := make(map[[2]int32]*plan.Expr)
+	node := builder.qry.Nodes[nodeID]
+
+	for i, childID := range node.Children {
+		newChildID, childProjMap := builder.removeSimpleProjections(childID, node.NodeType)
+		node.Children[i] = newChildID
+		for ref, expr := range childProjMap {
+			projMap[ref] = expr
+		}
+	}
+
+	removeProjectionsForExprList(node.ProjectList, projMap)
+	removeProjectionsForExprList(node.OnList, projMap)
+	removeProjectionsForExprList(node.FilterList, projMap)
+	removeProjectionsForExprList(node.GroupBy, projMap)
+	removeProjectionsForExprList(node.GroupingSet, projMap)
+	removeProjectionsForExprList(node.AggList, projMap)
+
+	for i := range node.OrderBy {
+		node.OrderBy[i].Expr = removeProjectionsForExpr(node.OrderBy[i].Expr, projMap)
+	}
+
+	if builder.canRemoveProject(parentType, node) {
+		allColRef := true
+		for _, proj := range node.ProjectList {
+			if _, ok := proj.Expr.(*plan.Expr_Col); !ok {
+				if _, ok := proj.Expr.(*plan.Expr_C); !ok {
+					allColRef = false
+					break
+				}
+			}
+		}
+
+		if allColRef {
+			tag := node.BindingTags[0]
+			for i, proj := range node.ProjectList {
+				projMap[[2]int32{tag, int32(i)}] = proj
+			}
+
+			nodeID = node.Children[0]
+		}
+	}
+
+	return nodeID, projMap
+}
+
+// FIXME: We should remove PROJECT node for more cases, but keep them now to avoid intricate issues.
+func (builder *QueryBuilder) canRemoveProject(parentType plan.Node_NodeType, node *plan.Node) bool {
+	if node.NodeType != plan.Node_PROJECT || node.Limit != nil || node.Offset != nil {
+		return false
+	}
+
+	if parentType == plan.Node_DISTINCT || parentType == plan.Node_UNKNOWN {
+		return false
+	}
+	if parentType == plan.Node_UNION || parentType == plan.Node_UNION_ALL {
+		return false
+	}
+	if parentType == plan.Node_MINUS || parentType == plan.Node_MINUS_ALL {
+		return false
+	}
+	if parentType == plan.Node_INTERSECT || parentType == plan.Node_INTERSECT_ALL {
+		return false
+	}
+	if parentType == plan.Node_FUNCTION_SCAN || parentType == plan.Node_EXTERNAL_FUNCTION {
+		return false
+	}
+
+	childType := builder.qry.Nodes[node.Children[0]].NodeType
+	if childType == plan.Node_VALUE_SCAN || childType == plan.Node_EXTERNAL_SCAN {
+		return false
+	}
+	if childType == plan.Node_FUNCTION_SCAN || childType == plan.Node_EXTERNAL_FUNCTION {
+		return false
+	}
+
+	return true
+}
+
+func removeProjectionsForExprList(exprList []*plan.Expr, projMap map[[2]int32]*plan.Expr) {
+	for i, expr := range exprList {
+		exprList[i] = removeProjectionsForExpr(expr, projMap)
+	}
+}
+
+func removeProjectionsForExpr(expr *plan.Expr, projMap map[[2]int32]*plan.Expr) *plan.Expr {
+	if expr == nil {
+		return nil
+	}
+
+	switch ne := expr.Expr.(type) {
+	case *plan.Expr_Col:
+		mapID := [2]int32{ne.Col.RelPos, ne.Col.ColPos}
+		if projExpr, ok := projMap[mapID]; ok {
+			return DeepCopyExpr(projExpr)
+		}
+
+	case *plan.Expr_F:
+		for i, arg := range ne.F.Args {
+			ne.F.Args[i] = removeProjectionsForExpr(arg, projMap)
+		}
+	}
+
+	return expr
+}
+
+func (builder *QueryBuilder) pushdownFilters(nodeID int32, filters []*plan.Expr, separateNonEquiConds bool) (int32, []*plan.Expr) {
+	node := builder.qry.Nodes[nodeID]
+
+	var canPushdown, cantPushdown []*plan.Expr
+
+	switch node.NodeType {
+	case plan.Node_AGG:
+		groupTag := node.BindingTags[0]
+		aggregateTag := node.BindingTags[1]
+
+		for _, filter := range filters {
+			if !containsTag(filter, aggregateTag) {
+				canPushdown = append(canPushdown, replaceColRefs(filter, groupTag, node.GroupBy))
+			} else {
+				cantPushdown = append(cantPushdown, filter)
+			}
+		}
+
+		childID, cantPushdownChild := builder.pushdownFilters(node.Children[0], canPushdown, separateNonEquiConds)
+
+		if len(cantPushdownChild) > 0 {
+			childID = builder.appendNode(&plan.Node{
+				NodeType:   plan.Node_FILTER,
+				Children:   []int32{node.Children[0]},
+				FilterList: cantPushdownChild,
+			}, nil)
+		}
+
+		node.Children[0] = childID
+
+	case plan.Node_FILTER:
+		canPushdown = filters
+		for _, filter := range node.FilterList {
+			canPushdown = append(canPushdown, splitPlanConjunction(applyDistributivity(builder.GetContext(), filter))...)
+		}
+
+		childID, cantPushdownChild := builder.pushdownFilters(node.Children[0], canPushdown, separateNonEquiConds)
+
+		if len(cantPushdownChild) > 0 {
+			node.Children[0] = childID
+			node.FilterList = cantPushdownChild
+		} else {
+			nodeID = childID
+		}
+
+	case plan.Node_JOIN:
+		leftTags := make(map[int32]*Binding)
+		for _, tag := range builder.enumerateTags(node.Children[0]) {
+			leftTags[tag] = nil
+		}
+
+		rightTags := make(map[int32]*Binding)
+		for _, tag := range builder.enumerateTags(node.Children[1]) {
+			rightTags[tag] = nil
+		}
+
+		var markTag int32
+		if node.JoinType == plan.Node_MARK {
+			markTag = node.BindingTags[0]
+		}
+
+		if node.JoinType == plan.Node_INNER {
+			for _, cond := range node.OnList {
+				filters = append(filters, splitPlanConjunction(applyDistributivity(builder.GetContext(), cond))...)
+			}
+
+			node.OnList = nil
+		}
+
+		var leftPushdown, rightPushdown []*plan.Expr
+		var turnInner bool
+
+		joinSides := make([]int8, len(filters))
+
+		for i, filter := range filters {
+			canTurnInner := true
+
+			joinSides[i] = getJoinSide(filter, leftTags, rightTags, markTag)
+			if f, ok := filter.Expr.(*plan.Expr_F); ok {
+				for _, arg := range f.F.Args {
+					if getJoinSide(arg, leftTags, rightTags, markTag) == JoinSideBoth {
+						canTurnInner = false
+						break
+					}
+				}
+			}
+
+			canTurnInner = canTurnInner && rejectsNull(filter, builder.compCtx.GetProcess())
+			leftOrRightJoin := (node.JoinType == plan.Node_LEFT && joinSides[i]&JoinSideRight != 0) || (node.JoinType == plan.Node_RIGHT && joinSides[i]&JoinSideLeft != 0)
+			if canTurnInner && leftOrRightJoin {
+				for _, cond := range node.OnList {
+					filters = append(filters, splitPlanConjunction(applyDistributivity(builder.GetContext(), cond))...)
+				}
+
+				node.JoinType = plan.Node_INNER
+				node.OnList = nil
+				turnInner = true
+
+				break
+			}
+
+			// TODO: FULL OUTER join should be handled here. However we don't have FULL OUTER join now.
+		}
+
+		if turnInner {
+			joinSides = make([]int8, len(filters))
+
+			for i, filter := range filters {
+				joinSides[i] = getJoinSide(filter, leftTags, rightTags, markTag)
+			}
+		} else if node.JoinType == plan.Node_LEFT {
+			var newOnList []*plan.Expr
+			for _, cond := range node.OnList {
+				conj := splitPlanConjunction(applyDistributivity(builder.GetContext(), cond))
+				for _, conjElem := range conj {
+					side := getJoinSide(conjElem, leftTags, rightTags, markTag)
+					if side&JoinSideLeft == 0 {
+						rightPushdown = append(rightPushdown, conjElem)
+					} else {
+						newOnList = append(newOnList, conjElem)
+					}
+				}
+			}
+
+			node.OnList = newOnList
+		}
+
+		if !separateNonEquiConds {
+			var extraFilters []*plan.Expr
+			for i, filter := range filters {
+				if joinSides[i] != JoinSideBoth {
+					continue
+				}
+				switch exprImpl := filter.Expr.(type) {
+				case *plan.Expr_F:
+					if exprImpl.F.Func.ObjName == "or" {
+						keys := checkDNF(filter)
+						for _, key := range keys {
+							extraFilter := walkThroughDNF(builder.GetContext(), filter, key)
+							if extraFilter != nil {
+								extraFilters = append(extraFilters, DeepCopyExpr(extraFilter))
+								joinSides = append(joinSides, getJoinSide(extraFilter, leftTags, rightTags, markTag))
+							}
+						}
+					}
+				}
+			}
+			filters = append(filters, extraFilters...)
+		}
+
+		for i, filter := range filters {
+			switch joinSides[i] {
+			case JoinSideNone:
+				if c, ok := filter.Expr.(*plan.Expr_C); ok {
+					if c, ok := c.C.Value.(*plan.Const_Bval); ok {
+						if c.Bval {
+							break
+						}
+					}
+				}
+
+				switch node.JoinType {
+				case plan.Node_INNER:
+					leftPushdown = append(leftPushdown, DeepCopyExpr(filter))
+					rightPushdown = append(rightPushdown, filter)
+
+				case plan.Node_LEFT, plan.Node_SEMI, plan.Node_ANTI, plan.Node_SINGLE, plan.Node_MARK:
+					leftPushdown = append(leftPushdown, filter)
+
+				default:
+					cantPushdown = append(cantPushdown, filter)
+				}
+
+			case JoinSideLeft:
+				if node.JoinType != plan.Node_OUTER && node.JoinType != plan.Node_RIGHT {
+					leftPushdown = append(leftPushdown, filter)
+				} else {
+					cantPushdown = append(cantPushdown, filter)
+				}
+
+			case JoinSideRight:
+				if node.JoinType == plan.Node_INNER || node.JoinType == plan.Node_RIGHT {
+					rightPushdown = append(rightPushdown, filter)
+				} else {
+					cantPushdown = append(cantPushdown, filter)
+				}
+
+			case JoinSideBoth:
+				if node.JoinType == plan.Node_INNER {
+					if separateNonEquiConds {
+						if f, ok := filter.Expr.(*plan.Expr_F); ok {
+							if f.F.Func.ObjName == "=" {
+								if getJoinSide(f.F.Args[0], leftTags, rightTags, markTag) != JoinSideBoth {
+									if getJoinSide(f.F.Args[1], leftTags, rightTags, markTag) != JoinSideBoth {
+										node.OnList = append(node.OnList, filter)
+										break
+									}
+								}
+							}
+						}
+					} else {
+						node.OnList = append(node.OnList, filter)
+						break
+					}
+				}
+
+				cantPushdown = append(cantPushdown, filter)
+
+			case JoinSideMark:
+				if tryMark, ok := filter.Expr.(*plan.Expr_Col); ok {
+					if tryMark.Col.RelPos == node.BindingTags[0] {
+						node.JoinType = plan.Node_SEMI
+						node.BindingTags = nil
+						break
+					}
+				} else if fExpr, ok := filter.Expr.(*plan.Expr_F); ok {
+					if filter.Typ.NotNullable && fExpr.F.Func.ObjName == "not" {
+						arg := fExpr.F.Args[0]
+						if tryMark, ok := arg.Expr.(*plan.Expr_Col); ok {
+							if tryMark.Col.RelPos == node.BindingTags[0] {
+								node.JoinType = plan.Node_ANTI
+								node.BindingTags = nil
+								break
+							}
+						}
+					}
+				}
+
+				cantPushdown = append(cantPushdown, filter)
+
+			default:
+				cantPushdown = append(cantPushdown, filter)
+			}
+		}
+
+		if node.JoinType == plan.Node_INNER {
+			//only inner join can deduce new predicate
+			builder.pushdownFilters(node.Children[0], predsDeduction(rightPushdown, node.OnList), separateNonEquiConds)
+			builder.pushdownFilters(node.Children[1], predsDeduction(leftPushdown, node.OnList), separateNonEquiConds)
+		}
+
+		childID, cantPushdownChild := builder.pushdownFilters(node.Children[0], leftPushdown, separateNonEquiConds)
+
+		if len(cantPushdownChild) > 0 {
+			childID = builder.appendNode(&plan.Node{
+				NodeType:   plan.Node_FILTER,
+				Children:   []int32{node.Children[0]},
+				FilterList: cantPushdownChild,
+			}, nil)
+		}
+
+		node.Children[0] = childID
+
+		childID, cantPushdownChild = builder.pushdownFilters(node.Children[1], rightPushdown, separateNonEquiConds)
+
+		if len(cantPushdownChild) > 0 {
+			childID = builder.appendNode(&plan.Node{
+				NodeType:   plan.Node_FILTER,
+				Children:   []int32{node.Children[1]},
+				FilterList: cantPushdownChild,
+			}, nil)
+		}
+
+		node.Children[1] = childID
+
+	case plan.Node_UNION, plan.Node_UNION_ALL, plan.Node_MINUS, plan.Node_MINUS_ALL, plan.Node_INTERSECT, plan.Node_INTERSECT_ALL:
+		leftChild := builder.qry.Nodes[node.Children[0]]
+		rightChild := builder.qry.Nodes[node.Children[1]]
+		var canPushDownRight []*plan.Expr
+
+		for _, filter := range filters {
+			canPushdown = append(canPushdown, replaceColRefsForSet(DeepCopyExpr(filter), leftChild.ProjectList))
+			canPushDownRight = append(canPushDownRight, replaceColRefsForSet(filter, rightChild.ProjectList))
+		}
+
+		childID, cantPushdownChild := builder.pushdownFilters(node.Children[0], canPushdown, separateNonEquiConds)
+		if len(cantPushdownChild) > 0 {
+			childID = builder.appendNode(&plan.Node{
+				NodeType:   plan.Node_FILTER,
+				Children:   []int32{node.Children[0]},
+				FilterList: cantPushdownChild,
+			}, nil)
+		}
+		node.Children[0] = childID
+
+		childID, cantPushdownChild = builder.pushdownFilters(node.Children[1], canPushDownRight, separateNonEquiConds)
+		if len(cantPushdownChild) > 0 {
+			childID = builder.appendNode(&plan.Node{
+				NodeType:   plan.Node_FILTER,
+				Children:   []int32{node.Children[1]},
+				FilterList: cantPushdownChild,
+			}, nil)
+		}
+		node.Children[1] = childID
+
+	case plan.Node_PROJECT:
+		child := builder.qry.Nodes[node.Children[0]]
+		if (child.NodeType == plan.Node_VALUE_SCAN || child.NodeType == plan.Node_EXTERNAL_SCAN) && child.RowsetData == nil {
+			cantPushdown = filters
+			break
+		}
+
+		projectTag := node.BindingTags[0]
+
+		for _, filter := range filters {
+			canPushdown = append(canPushdown, replaceColRefs(filter, projectTag, node.ProjectList))
+		}
+
+		childID, cantPushdownChild := builder.pushdownFilters(node.Children[0], canPushdown, separateNonEquiConds)
+
+		if len(cantPushdownChild) > 0 {
+			childID = builder.appendNode(&plan.Node{
+				NodeType:   plan.Node_FILTER,
+				Children:   []int32{node.Children[0]},
+				FilterList: cantPushdownChild,
+			}, nil)
+		}
+
+		node.Children[0] = childID
+
+	case plan.Node_TABLE_SCAN, plan.Node_EXTERNAL_SCAN:
+		node.FilterList = append(node.FilterList, filters...)
+	case plan.Node_FUNCTION_SCAN:
+		node.FilterList = append(node.FilterList, filters...)
+		childId := node.Children[0]
+		childId, err := builder.pushdownFilters(childId, nil, separateNonEquiConds)
+		if err != nil {
+			return 0, err
+		}
+		node.Children[0] = childId
+
+	default:
+		if len(node.Children) > 0 {
+			childID, cantPushdownChild := builder.pushdownFilters(node.Children[0], filters, separateNonEquiConds)
+
+			if len(cantPushdownChild) > 0 {
+				childID = builder.appendNode(&plan.Node{
+					NodeType:   plan.Node_FILTER,
+					Children:   []int32{node.Children[0]},
+					FilterList: cantPushdownChild,
+				}, nil)
+			}
+
+			node.Children[0] = childID
+		} else {
+			cantPushdown = filters
+		}
+	}
+
+	return nodeID, cantPushdown
+}
